touched install

+++ utility/mysqld_purge.sh
@@ -1,4 +1,12 @@
 :
+if [ $# -ne 1 ]
+then
+	echo "Usage: $0 version.release" 1>&2
+	exit 1
+fi
+
+version=$1
+
 sudo service mysqld stop
 sudo killall -9 mysql
 sudo killall -9 mysqld
@@ -8,8 +16,8 @@ sudo apt-get autoclean
 sudo deluser mysql
 sudo delgroup mysql
 sudo rm -rf /var/lib/mysql
-sudo apt-get purge mysql-server-core-5.5
-sudo apt-get purge mysql-client-core-5.5
+sudo apt-get purge mysql-server-core-$version
+sudo apt-get purge mysql-client-core-$version
 sudo rm -rf /var/log/mysql
 sudo rm -rf /etc/mysql
 sudo rm -rf /var/log/mysql
